Move out Osd PackedPatchTable implementation, hide implementation details.
