+++ src/render/mesh_subdivision.cpp
@@ -310,6 +310,8 @@ class OsdData {
     return min(isolation, 10);
   }
 
+  void pack(PackedPatchTable* packed_patch_table, int offset = 0) const;
+
   friend struct OsdPatch;
   friend class Mesh;
 };
@@ -363,6 +365,207 @@ struct OsdPatch final : Patch {
     }
   }
 };
+
+struct PatchMapQuadNode {
+    /* sets all the children to point to the patch of index */
+    void set_child(int index)
+    {
+        for (int i = 0; i < 4; i++) {
+            children[i] = index | PATCH_MAP_NODE_IS_SET | PATCH_MAP_NODE_IS_LEAF;
+        }
+    }
+
+    /* sets the child in quadrant to point to the node or patch of the given index */
+    void set_child(unsigned char quadrant, int index, bool is_leaf = true)
+    {
+        assert(quadrant < 4);
+        children[quadrant] = index | PATCH_MAP_NODE_IS_SET | (is_leaf ? PATCH_MAP_NODE_IS_LEAF : 0);
+    }
+
+    uint children[4];
+};
+
+template<class T> static int resolve_quadrant(T &median, T &u, T &v)
+        {
+    int quadrant = -1;
+
+    if (u < median) {
+        if (v < median) {
+            quadrant = 0;
+        }
+        else {
+            quadrant = 1;
+            v -= median;
+        }
+    }
+    else {
+        if (v < median) {
+            quadrant = 3;
+        }
+        else {
+            quadrant = 2;
+            v -= median;
+        }
+        u -= median;
+    }
+
+    return quadrant;
+}
+
+static void build_patch_map(PackedPatchTable &table,
+                            const Far::PatchTable *patch_table,
+                            int offset)
+                            {
+    int num_faces = 0;
+
+    for (int array = 0; array < table.num_arrays; array++) {
+        Far::ConstPatchParamArray params = patch_table->GetPatchParams(array);
+
+        for (int j = 0; j < patch_table->GetNumPatches(array); j++) {
+            num_faces = max(num_faces, (int)params[j].GetFaceId());
+        }
+    }
+    num_faces++;
+
+    vector<PatchMapQuadNode> quadtree;
+    quadtree.reserve(num_faces + table.num_patches);
+    quadtree.resize(num_faces);
+
+    /* adjust offsets to make indices relative to the table */
+    int handle_index = -(table.num_patches * PATCH_HANDLE_SIZE);
+    offset += table.total_size();
+
+    /* populate the quadtree from the FarPatchArrays sub-patches */
+    for (int array = 0; array < table.num_arrays; array++) {
+        Far::ConstPatchParamArray params = patch_table->GetPatchParams(array);
+
+        for (int i = 0; i < patch_table->GetNumPatches(array);
+        i++, handle_index += PATCH_HANDLE_SIZE) {
+            const Far::PatchParam &param = params[i];
+            unsigned short depth = param.GetDepth();
+
+            PatchMapQuadNode *node = &quadtree[params[i].GetFaceId()];
+
+            if (depth == (param.NonQuadRoot() ? 1 : 0)) {
+                /* special case : regular BSpline face w/ no sub-patches */
+                node->set_child(handle_index + offset);
+                continue;
+            }
+
+            int u = param.GetU();
+            int v = param.GetV();
+            int pdepth = param.NonQuadRoot() ? depth - 2 : depth - 1;
+            int half = 1 << pdepth;
+
+            for (int j = 0; j < depth; j++) {
+                int delta = half >> 1;
+
+                int quadrant = resolve_quadrant(half, u, v);
+                assert(quadrant >= 0);
+
+                half = delta;
+
+                if (j == pdepth) {
+                    /* we have reached the depth of the sub-patch : add a leaf */
+                    assert(!(node->children[quadrant] & PATCH_MAP_NODE_IS_SET));
+                    node->set_child(quadrant, handle_index + offset, true);
+                    break;
+                }
+                else {
+                    /* travel down the child node of the corresponding quadrant */
+                    if (!(node->children[quadrant] & PATCH_MAP_NODE_IS_SET)) {
+                        /* create a new branch in the quadrant */
+                        quadtree.push_back(PatchMapQuadNode());
+
+                        int idx = (int)quadtree.size() - 1;
+                        node->set_child(quadrant, idx * 4 + offset, false);
+
+                        node = &quadtree[idx];
+                    }
+                    else {
+                        /* travel down an existing branch */
+                        uint idx = node->children[quadrant] & PATCH_MAP_NODE_INDEX_MASK;
+                        node = &(quadtree[(idx - offset) / 4]);
+                    }
+                }
+            }
+        }
+    }
+
+    /* copy into table */
+    assert(table.table.size() == table.total_size());
+    uint map_offset = table.total_size();
+
+    table.num_nodes = quadtree.size() * 4;
+    table.table.resize(table.total_size());
+
+    uint *data = &table.table[map_offset];
+
+    for (int i = 0; i < quadtree.size(); i++) {
+        for (int j = 0; j < 4; j++) {
+            assert(quadtree[i].children[j] & PATCH_MAP_NODE_IS_SET);
+            *(data++) = quadtree[i].children[j];
+        }
+    }
+}
+
+void OsdData::pack(PackedPatchTable* packed_patch_table, int offset) const
+{
+    packed_patch_table->num_arrays = 0;
+    packed_patch_table->num_patches = 0;
+    packed_patch_table->num_indices = 0;
+    packed_patch_table->num_nodes = 0;
+
+    packed_patch_table->num_arrays = patch_table->GetNumPatchArrays();
+
+    for (int i = 0; i < packed_patch_table->num_arrays; i++) {
+        int patches = patch_table->GetNumPatches(i);
+        int num_control = patch_table->GetPatchArrayDescriptor(i).GetNumControlVertices();
+
+        packed_patch_table->num_patches += patches;
+        packed_patch_table->num_indices += patches * num_control;
+    }
+
+    packed_patch_table->table.resize(packed_patch_table->total_size());
+    uint *data = packed_patch_table->table.data();
+
+    uint *array = data;
+    uint *index = array + packed_patch_table->num_arrays * PATCH_ARRAY_SIZE;
+    uint *param = index + packed_patch_table->num_indices;
+    uint *handle = param + packed_patch_table->num_patches * PATCH_PARAM_SIZE;
+
+    uint current_param = 0;
+
+    for (int i = 0; i < packed_patch_table->num_arrays; i++) {
+        *(array++) = patch_table->GetPatchArrayDescriptor(i).GetType();
+        *(array++) = patch_table->GetNumPatches(i);
+        *(array++) = (index - data) + offset;
+        *(array++) = (param - data) + offset;
+
+        Far::ConstIndexArray indices = patch_table->GetPatchArrayVertices(i);
+
+        for (int j = 0; j < indices.size(); j++) {
+            *(index++) = indices[j];
+        }
+
+        const Far::PatchParamTable &param_table = patch_table->GetPatchParamTable();
+
+        int num_control = patch_table->GetPatchArrayDescriptor(i).GetNumControlVertices();
+        int patches = patch_table->GetNumPatches(i);
+
+        for (int j = 0; j < patches; j++, current_param++) {
+            *(param++) = param_table[current_param].field0;
+            *(param++) = param_table[current_param].field1;
+
+            *(handle++) = (array - data) - PATCH_ARRAY_SIZE + offset;
+            *(handle++) = (param - data) - PATCH_PARAM_SIZE + offset;
+            *(handle++) = j * num_control;
+        }
+    }
+
+    build_patch_map(*packed_patch_table, patch_table, offset);
+}
+
 
 #endif
 
@@ -633,7 +836,7 @@ void Mesh::tessellate(DiagSplit *split)
   if (need_packed_patch_table) {
     delete patch_table;
     patch_table = new PackedPatchTable;
-    patch_table->pack(osd_data.patch_table);
+    osd_data.pack(patch_table);
   }
 #endif
 }
